+++ CHANGELOG.md
@@ -33,6 +33,7 @@ All user visible changes to this project will be documented in this file. This p
 - Unexpected audio category on `setOutputAudioId` call on [iOS]. ([#146])
 - Race condition bug on `setOutputAudioId` call on [Android]. ([#146])
 - Race condition bug on input/output device switch on desktop platforms. ([#151])
+- `RtpReceiver` use after free on [Android]. ([#165])
 
 [#137]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/137
 [#139]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/139
@@ -47,6 +48,7 @@ All user visible changes to this project will be documented in this file. This p
 [#156]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/156
 [#162]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/162
 [#164]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/164
+[#165]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/165
 [`dart:html`]: https://dart.dev/libraries/dart-html
 [`package:web`]: https://pub.dev/packages/web
 [126.0.6478.182-r2]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/126.0.6478.182-r2